requirements.txt
# Core FastAPI dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# PII Detection and NLP
presidio-analyzer>=2.2.0
presidio-anonymizer>=2.2.0
spacy>=3.7.0
phonenumbers>=8.13.0

# Database and Storage
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0

# Caching and Rate Limiting
redis>=5.0.0

# HTTP and WebSocket
httpx>=0.25.0
websockets>=12.0

# Configuration and Utilities
python-multipart>=0.0.6
pyyaml>=6.0.0
pytesseract>=0.3.10
Pillow>=10.0.0

# Monitoring and Metrics
prometheus-client>=0.19.0
